004_airflow_dbt_retail_project/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

apache-airflow-providers-google==10.14.0
soda-core-bigquery==3.0.45

google-cloud-bigquery-storage
# opentelemetry-instrumentation
